+++ core-ng-mongo/src/main/java/core/framework/mongo/Mongo.java
@@ -15,6 +15,8 @@ public interface Mongo {
     // refer to com.mongodb.client.model.Indexes for building keys
     void createIndex(String collection, Bson keys, IndexOptions options);
 
+    void dropIndex(String collection, Bson keys);
+
     void dropCollection(String collection);
 
     Document runCommand(Bson command);